+++ src/main/java/main/arbitrage/domain/buyOrder/entity/BuyOrder.java
@@ -49,25 +49,25 @@ public class BuyOrder {
     private ExchangeRate exchangeRate;
 
     @Column(name = "premium", nullable = false, columnDefinition = "REAL")
-    private float premium;
+    private double premium;
 
     @Column(name = "upbit_price", nullable = false, columnDefinition = "DOUBLE PRECISION")
     private double upbitPrice;
 
     @Column(name = "upbit_quantity", nullable = false, columnDefinition = "REAL")
-    private float upbitQuantity;
+    private double upbitQuantity;
 
     @Column(name = "upbit_commission", nullable = false, columnDefinition = "REAL")
-    private float upbitCommission;
+    private double upbitCommission;
 
     @Column(name = "binance_price", nullable = false, columnDefinition = "REAL")
-    private float binancePrice;
+    private double binancePrice;
 
     @Column(name = "binance_quantity", nullable = false, columnDefinition = "REAL")
-    private float binanceQuantity;
+    private double binanceQuantity;
 
     @Column(name = "binance_commission", nullable = false, columnDefinition = "REAL")
-    private float binanceCommission;
+    private double binanceCommission;
 
     @Column(name = "is_maker", columnDefinition = "BOOLEAN DEFAULT FALSE")
     private boolean isMaker;
@@ -87,9 +87,9 @@ public class BuyOrder {
     private List<SellOrder> sellOrders = new ArrayList<>();
 
     @Builder
-    public BuyOrder(User user, Symbol symbol, ExchangeRate exchangeRate, float premium,
-            double upbitPrice, float upbitQuantity, float upbitCommission, float binancePrice,
-            float binanceQuantity, float binanceCommission, boolean isMaker, boolean isClose) {
+    public BuyOrder(User user, Symbol symbol, ExchangeRate exchangeRate, double premium,
+            double upbitPrice, double upbitQuantity, double upbitCommission, double binancePrice,
+            double binanceQuantity, double binanceCommission, boolean isMaker, boolean isClose) {
         this.user = user;
         this.symbol = symbol;
         this.exchangeRate = exchangeRate;
