+++ src/main/java/main/arbitrage/infrastructure/crawler/UsdToKrwCrawler.java
@@ -13,7 +13,7 @@ public class UsdToKrwCrawler {
             "https://www.google.com/finance/quote/USD-KRW";
     private static final String CURRENT_SELECTOR = "div.YMlKec.fxKbKc"; // 환율 셀렉터
 
-    public float craw() {
+    public double craw() {
         try {
             Document doc = Jsoup.connect(GOOGLE_FINANCE_USD_TO_KRW_URL).get();
 
@@ -22,7 +22,7 @@ public class UsdToKrwCrawler {
             if (rateToString.isEmpty())
                 return 0;
 
-            return Float.parseFloat(rateToString.replace(",", ""));
+            return Double.parseDouble(rateToString.replace(",", ""));
         } catch (IOException e) {
             // Jsoup로 crawling을 했을때 반환값이 0일경우 아래와 같은 Error Message가 발생함.
             // 에러 확인 후 코드 추가 -> 아직 해당 에러 재발생하지 않음. 추후에 수정.
